refactor: move graphsession pkg to graph package
