+++ graph/graphsession/graph_session.go
